+++ demo/index.html
@@ -49,9 +49,8 @@
     </div>
     <script type="text/javascript" src="http://localhost:8080/build/umd/core.development.js"></script>
     <script type="text/javascript" src="http://localhost:8080/build/umd/networks.dfp.development.js"></script>
-    <script type="text/javascript" src="http://localhost:8080/build/umd/plugins.logging.development.js"></script>
+    <script type="text/javascript" src="http://localhost:8080/build/umd/plugins.developertools.development.js"></script>
     <script type="text/javascript" src="http://localhost:8080/build/umd/plugins.autorender.development.js"></script>
-    <script type="text/javascript" src="http://localhost:8080/build/umd/plugins.debug.development.js"></script>
     <script type="text/javascript" src="http://localhost:8080/build/umd/plugins.sticky.development.js"></script>
     <script type="text/javascript" src="http://localhost:8080/build/umd/plugins.autorefresh.development.js"></script>
     <script type="text/javascript" src="http://localhost:8080/build/umd/plugins.responsive.development.js"></script>
@@ -63,7 +62,7 @@
           AdJS.Plugins.Logging,
           AdJS.Plugins.AutoRender,
           AdJS.Plugins.AutoRefresh,
-          AdJS.Plugins.Debug,
+          ADJS.Plugins.DeveloperTools,
           AdJS.Plugins.Sticky,
           AdJS.Plugins.Responsive,
         ],
