--- docs/debug-tools.md
@@ -1,4 +1,0 @@
-# Debugging
-
-An advantage with using Ad.js is the debug interface. At any point you may call AdJS.DEBUG() in the console and as long as your page includes the Debug module, information on all details of the ad calls/code/etc will appear on screen where the ad would normally render. TODO: elaborate and screenshots
-
